Cleaned up custom user request type
