+++ server/src/routes/listings.route.ts
@@ -4,8 +4,7 @@ import { sellerAuth } from "../middlewares/sellerAuth";
 import {
   createListing,
   getAllListings,
-  getUserListingById,
-  // getListingsByUserId,
+  getListingById,
   updateListing,
   deleteListing,
 } from "../controllers/listings.controller";
@@ -18,8 +17,7 @@ router.post("/", createListing);
 // GET /listings/:id - Get a listing by ID
 // PATCH /listings/:id - Update a listing by ID
 // DELETE /listings/:id - Delete a listing by ID
-router.get("/:id", sellerAuth, getUserListingById);
-// router.get("/user/:id", getListingsByUserId);
+router.get("/:id", sellerAuth, getListingById);
 router.patch("/:id", sellerAuth, updateListing);
 router.delete("/:id", sellerAuth, deleteListing);
