--- server/src/types/RequestObject.ts
@@ -1,6 +1,0 @@
-export default interface RequestObject {
-  userID: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-}
