+++ server/src/types/UserRequest.ts
@@ -1,0 +1,10 @@
+import { Request } from "express";
+
+export interface UserRequest extends Request {
+  user: {
+    userID: string;
+    firstName: string;
+    lastName: string;
+    email: string;
+  };
+}
