--- server/src/types/request.ts
@@ -1,24 +1,0 @@
-import { Request } from "express";
-// import { returnObject } from "./return.object";
-
-// declare global {
-//   namespace Express {
-//     interface Request {
-//       user?: {
-//         userID: string;
-//         firstName: string;
-//         lastName: string;
-//         email: string;
-//       };
-//     }
-//   }
-// }
-
-export interface UserRequest extends Request {
-  user: {
-    userID: string;
-    firstName: string;
-    lastName: string;
-    email: string;
-  };
-}
